Fix deletion

+++ clive/__private/storage/service.py
@@ -4,8 +4,7 @@ import contextlib
 import re
 import shutil
 from dataclasses import dataclass
-from pathlib import Path
-from typing import TYPE_CHECKING, Final
+from typing import TYPE_CHECKING, Final, Literal
 
 from clive.__private.core.commands.abc.command_encryption import CommandRequiresUnlockedEncryptionWalletError
 from clive.__private.core.commands.decrypt import CommandDecryptError
@@ -19,6 +18,7 @@ from clive.exceptions import CliveError
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
+    from pathlib import Path
 
     from clive.__private.core.encryption import EncryptionService
     from clive.__private.core.profile import Profile
@@ -76,14 +76,18 @@ class PersistentStorageService:
     FIRST_REVISION: Final[str] = "c600278a"
 
     @dataclass(frozen=True)
-    class ProfileNameModel:
-        name: str
-        model_cls: type[ProfileStorageBase]
+    class ProfileFileInfo:
+        profile_name: str
+        path: Path
+        model_cls: type[ProfileStorageBase] | None = None
+
+        @property
+        def is_backup(self) -> bool:
+            return self.model_cls is None
 
         def version(self) -> int:
+            assert self.model_cls is not None, "Can't get version for backup file."
             return self.model_cls.get_this_version()
-
-    type ProfileNameModelToPath = dict[ProfileNameModel, Path]
 
     def __init__(self, encryption_service: EncryptionService) -> None:
         self._encryption_service = encryption_service
@@ -149,22 +153,22 @@ class PersistentStorageService:
             ProfileDoesNotExistsError: If profile with given name does not exist, it could not be removed.
             MultipleProfileVersionsError: If multiple versions / back-ups of profile exist and force is False.
         """
-        if not force:
-            cls._raise_if_profile_not_stored(profile_name)
+        filepaths = cls._get_filepaths(profile_name, file_type="all", include_impossible_to_load=True)
+        num_of_files_to_remove = len(filepaths)
+        if num_of_files_to_remove == 0:
+            raise ProfileDoesNotExistsError(profile_name)
+        if num_of_files_to_remove > 1 and not force:
+            raise MultipleProfileVersionsError(profile_name)
 
         cls._delete_legacy_profile_data(profile_name)
-
         profile_dir = cls.get_profile_directory(profile_name)
-        num_of_files_in_profile_dir = len(list(profile_dir.glob("*")))
-        if num_of_files_in_profile_dir > 1 and not force:
-            raise MultipleProfileVersionsError(profile_name)
-        if profile_dir.exists():  # we might store only legacy version of the profile
+        if profile_dir.exists():  # we can store only a legacy version of the profile
             shutil.rmtree(profile_dir)
 
     @classmethod
     def list_stored_profile_names(cls) -> list[str]:
         """List all stored profile names sorted alphabetically including older storage versions."""
-        profile_names = {key.name for key in cls._get_filepaths()}
+        profile_names = {filepath.profile_name for filepath in cls._get_filepaths()}
         return sorted(profile_names)
 
     @classmethod
@@ -226,36 +230,63 @@ class PersistentStorageService:
 
     @classmethod
     def _get_profile_filepath_to_read(cls, profile_name: str) -> Path | None:
-        filepaths = cls._get_filepaths()
-
-        candidates = [item for item in filepaths if item.name == profile_name]
-
-        if not candidates:
+        filepaths = cls._get_filepaths(profile_name)
+        if not filepaths:
             return None
 
         # Pick the path with the highest version
-        key = max(candidates, key=lambda item: item.version())
-        return filepaths[key]
+        highest_version = max(filepaths, key=lambda item: item.version())
+        return highest_version.path
 
     @classmethod
-    def _get_filepaths(cls) -> ProfileNameModelToPath:
+    def _get_filepaths(
+        cls,
+        profile_name: str | None = None,
+        file_type: Literal["profile", "backup", "all"] = "profile",
+        *,
+        include_impossible_to_load: bool = False,
+    ) -> set[PersistentStorageService.ProfileFileInfo]:
         """
         Retrieve file paths of profiles stored on the disk.
 
-        It will retrieve only profiles that we have a corresponding version of model for.
+        If include_impossible_to_load=False, it will retrieve only profiles that
+        we have a corresponding version of model for.
         It means any newer versions won't be picked up. (Like we're on v2, but there is v3.profile)
-        """
-        path_extension = cls.PROFILE_FILENAME_SUFFIX
-        storage_dir = cls._get_storage_directory()
-        paths: PersistentStorageService.ProfileNameModelToPath = {}
 
-        for path in storage_dir.glob(f"*/*{path_extension}"):
-            if cls._is_model_cls_for_versioned_profile_file_available(path):
-                name_model = cls.ProfileNameModel(
-                    name=cls._profile_name_from_path(path),
-                    model_cls=cls._model_cls_from_path(path),
-                )
-                paths[name_model] = path
+        Args:
+        ----
+            profile_name: If given, only profiles with this name will be returned. If None, return all profiles.
+            file_type: Determine which type of files to look for.
+            include_impossible_to_load: If True, it will return profiles even we could not load them.
+        """
+        file_type_extension_mapping: dict[Literal["profile", "backup", "all"], str] = {
+            "profile": cls.PROFILE_FILENAME_SUFFIX,
+            "backup": cls.BACKUP_FILENAME_SUFFIX,
+            "all": ".*",
+        }
+        path_suffix_to_search = file_type_extension_mapping[file_type]
+        storage_dir = cls._get_storage_directory()
+        paths: set[PersistentStorageService.ProfileFileInfo] = set()
+
+        for path in storage_dir.glob(f"*/*{path_suffix_to_search}"):
+            path_suffix = path.suffix
+            if (
+                not include_impossible_to_load
+                and path_suffix == cls.PROFILE_FILENAME_SUFFIX
+                and not cls._is_model_cls_for_versioned_profile_file_available(path)
+            ):
+                continue
+
+            profile_name_from_path = cls._profile_name_from_path(path)
+            if profile_name is not None and profile_name != profile_name_from_path:
+                continue
+
+            name_model = cls.ProfileFileInfo(
+                profile_name=profile_name_from_path,
+                model_cls=cls._model_cls_from_path(path) if path_suffix == cls.PROFILE_FILENAME_SUFFIX else None,
+                path=path,
+            )
+            paths.add(name_model)
         return paths
 
     async def _save_profile_model(self, profile_model: ProfileStorageModel) -> None:
